requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements-all.txt requirements-all.in
#
alembic==1.16.2
    # via mlflow
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
babel==2.17.0
    # via mkdocs-material
blinker==1.9.0
    # via flask
boto3==1.39.3
    # via metaflow
botocore==1.39.3
    # via
    #   boto3
    #   s3transfer
build==1.2.2.post1
    # via pip-tools
certifi==2025.6.15
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
click==8.1.7
    # via
    #   -r requirements-all.in
    #   databricks-cli
    #   flask
    #   mkdocs
    #   mlflow
    #   pip-tools
cloudpickle==3.1.1
    # via mlflow
colorama==0.4.6
    # via mkdocs-material
contourpy==1.3.2
    # via matplotlib
coverage[toml]==7.9.2
    # via pytest-cov
cramjam==2.10.0
    # via fastparquet
cycler==0.12.1
    # via matplotlib
databricks-cli==0.18.0
    # via mlflow
distlib==0.3.9
    # via virtualenv
docker==6.1.3
    # via mlflow
entrypoints==0.4
    # via mlflow
fastparquet==2023.10.0
    # via -r requirements-all.in
filelock==3.18.0
    # via virtualenv
flask==3.1.1
    # via mlflow
fonttools==4.58.5
    # via matplotlib
fsspec==2025.5.1
    # via fastparquet
ghp-import==2.1.0
    # via mkdocs
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via mlflow
gunicorn==21.2.0
    # via mlflow
identify==2.6.12
    # via pre-commit
idna==3.10
    # via requests
importlib-metadata==7.2.1
    # via mlflow
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   mkdocs
    #   mkdocs-material
    #   mlflow
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.5.1
    # via scikit-learn
jsonschema==4.20.0
    # via -r requirements-all.in
jsonschema-specifications==2025.4.1
    # via jsonschema
kiwisolver==1.4.8
    # via matplotlib
mako==1.3.10
    # via alembic
markdown==3.8.2
    # via
    #   mkdocs
    #   mkdocs-material
    #   mlflow
    #   pymdown-extensions
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   mkdocs
    #   werkzeug
matplotlib==3.10.3
    # via mlflow
mdurl==0.1.2
    # via markdown-it-py
mergedeep==1.3.4
    # via mkdocs
metaflow==2.11.0
    # via -r requirements-all.in
mkdocs==1.5.3
    # via
    #   -r requirements-all.in
    #   mkdocs-material
mkdocs-material==9.5.0
    # via -r requirements-all.in
mkdocs-material-extensions==1.3.1
    # via mkdocs-material
mlflow==2.9.0
    # via -r requirements-all.in
mypy==1.7.0
    # via -r requirements-all.in
mypy-extensions==1.1.0
    # via mypy
nodeenv==1.9.1
    # via pre-commit
numpy==1.26.4
    # via
    #   -r requirements-all.in
    #   contourpy
    #   fastparquet
    #   matplotlib
    #   mlflow
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
oauthlib==3.3.1
    # via databricks-cli
packaging==23.2
    # via
    #   build
    #   docker
    #   fastparquet
    #   gunicorn
    #   matplotlib
    #   mkdocs
    #   mlflow
    #   pytest
paginate==0.5.7
    # via mkdocs-material
pandas==2.0.3
    # via
    #   -r requirements-all.in
    #   fastparquet
    #   mlflow
pathspec==0.12.1
    # via mkdocs
pillow==11.3.0
    # via matplotlib
pip-tools==7.3.0
    # via -r requirements-all.in
platformdirs==4.3.8
    # via
    #   mkdocs
    #   virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==3.6.0
    # via -r requirements-all.in
protobuf==4.25.8
    # via mlflow
pyarrow==14.0.2
    # via
    #   -r requirements-all.in
    #   mlflow
pydantic==2.5.0
    # via -r requirements-all.in
pydantic-core==2.14.1
    # via pydantic
pygments==2.19.2
    # via
    #   mkdocs-material
    #   rich
pyjwt==2.10.1
    # via databricks-cli
pymdown-extensions==10.16
    # via mkdocs-material
pyparsing==3.2.3
    # via matplotlib
pyproject-hooks==1.2.0
    # via build
pytest==7.4.3
    # via
    #   -r requirements-all.in
    #   pytest-cov
    #   pytest-mock
pytest-cov==4.1.0
    # via -r requirements-all.in
pytest-mock==3.12.0
    # via -r requirements-all.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   ghp-import
    #   matplotlib
    #   pandas
python-dotenv==1.0.0
    # via -r requirements-all.in
pytz==2023.4
    # via
    #   mlflow
    #   pandas
pyyaml==6.0.1
    # via
    #   -r requirements-all.in
    #   mkdocs
    #   mlflow
    #   pre-commit
    #   pymdown-extensions
    #   pyyaml-env-tag
pyyaml-env-tag==1.1
    # via mkdocs
querystring-parser==1.2.4
    # via mlflow
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via mkdocs-material
requests==2.32.4
    # via
    #   databricks-cli
    #   docker
    #   metaflow
    #   mkdocs-material
    #   mlflow
rich==13.7.0
    # via -r requirements-all.in
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
ruff==0.1.8
    # via -r requirements-all.in
s3transfer==0.13.0
    # via boto3
scikit-learn==1.4.0
    # via
    #   -r requirements-all.in
    #   mlflow
scipy==1.16.0
    # via
    #   mlflow
    #   scikit-learn
six==1.17.0
    # via
    #   databricks-cli
    #   python-dateutil
    #   querystring-parser
smmap==5.0.2
    # via gitdb
sqlalchemy==2.0.41
    # via
    #   alembic
    #   mlflow
sqlparse==0.5.3
    # via mlflow
tabulate==0.9.0
    # via databricks-cli
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.66.1
    # via -r requirements-all.in
typing-extensions==4.14.1
    # via
    #   alembic
    #   mypy
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   botocore
    #   databricks-cli
    #   docker
    #   requests
virtualenv==20.31.2
    # via pre-commit
watchdog==6.0.0
    # via mkdocs
websocket-client==1.8.0
    # via docker
werkzeug==3.1.3
    # via flask
wheel==0.45.1
    # via pip-tools
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
